Added carbon example folder
