+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/Gantt/BasicGanttCharts/Actions.jsx
@@ -1,7 +1,9 @@
 import React from 'react';
-import GanttChart from '@cerner/terra-graphs/lib/components/Gantt/GanttChart';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/Gantt/GanttChart.module.scss';
 import getGanttDefaultConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Gantt/ganttDefault';
 import loadTracks from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Gantt/loadTracks';
 import actionsData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Gantt/actionsData';
@@ -11,19 +13,22 @@ Please refer to the documentation below to see the graphConfig and data objects
 */
 const graphConfig = utils.deepClone(getGanttDefaultConfig('#ganttActions'));
 graphConfig.showActionLegend = true;
-graphConfig.removeContainerPadding = true;
-graphConfig.legendPadding = {
-  left: 2.5,
-  right: 2.5,
-  top: 2.5,
-  bottom: 2.5,
-};
 const actions = utils.deepClone(actionsData);
 const dataset = loadTracks([], [], [], actions, 2);
 
-export default () => (
-  <React.Fragment>
-    <div id="tooltip" className="initial-tooltip" />
-    <GanttChart graphID="ganttActions" graphConfig={graphConfig} dataset={dataset} />
-  </React.Fragment>
-);
+const GanttExample = () => {
+  React.useEffect(() => {
+    const graph = Carbon.api.gantt(graphConfig);
+    dataset.forEach((data) => {
+      graph.loadContent(data);
+    });
+  }, []);
+
+  return (
+    <React.Fragment>
+      <div id="tooltip" className="initial-tooltip" />
+      <div id="ganttActions" />
+    </React.Fragment>
+  );
+};
+export default GanttExample;
