+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/Gantt/BasicGanttCharts/BarTypes.jsx
@@ -1,9 +1,11 @@
 import React from 'react';
-import GanttChart from '@cerner/terra-graphs/lib/components/Gantt/GanttChart';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/Gantt/GanttChart.module.scss';
 import getGanttDefaultConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Gantt/ganttDefault';
 import contentArray from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Gantt/barTypesData';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
@@ -14,9 +16,20 @@ graphConfig.dateline = [];
 
 const dataset = utils.deepClone(contentArray);
 
-export default () => (
-  <React.Fragment>
-    <div id="tooltip" className="initial-tooltip" />
-    <GanttChart graphID="ganttBarType" graphConfig={graphConfig} dataset={dataset} />
-  </React.Fragment>
-);
+const GanttExample = () => {
+  React.useEffect(() => {
+    const graph = Carbon.api.gantt(graphConfig);
+    dataset.forEach((data) => {
+      graph.loadContent(data);
+    });
+  }, []);
+
+  return (
+    <React.Fragment>
+      <div id="tooltip" className="initial-tooltip" />
+      <div id="ganttBarType" />
+    </React.Fragment>
+  );
+};
+
+export default GanttExample;
