+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/Gantt/BasicGanttCharts/Eventline.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
-import GanttChart from '@cerner/terra-graphs/lib/components/Gantt/GanttChart';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/Gantt/GanttChart.module.scss';
 import getGanttDefaultConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Gantt/ganttDefault';
 import loadTracks from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Gantt/loadTracks';
 import tasksData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Gantt/tasksData';
@@ -30,9 +31,20 @@ graphConfig.eventline = [
 graphConfig.showActionLegend = false;
 const dataset = loadTracks(tasksData, [], [], [], 4);
 
-export default () => (
-  <React.Fragment>
-    <div id="tooltip" className="initial-tooltip" />
-    <GanttChart graphID="ganttEventline" graphConfig={graphConfig} dataset={dataset} />
-  </React.Fragment>
-);
+const GanttExample = () => {
+  React.useEffect(() => {
+    const graph = Carbon.api.gantt(graphConfig);
+    dataset.forEach((data) => {
+      graph.loadContent(data);
+    });
+  }, []);
+
+  return (
+    <React.Fragment>
+      <div id="tooltip" className="initial-tooltip" />
+      <div id="ganttEventline" />
+    </React.Fragment>
+  );
+};
+
+export default GanttExample;
