+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/Gantt/Reflow/DynamicData.jsx
@@ -1,23 +1,36 @@
 import React from 'react';
-import GanttChart from '@cerner/terra-graphs/lib/components/Gantt/GanttChart';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 import Button from 'terra-button/lib/Button';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/Gantt/GanttChart.module.scss';
 import getGanttPanningConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Gantt/ganttPanning';
-import graphData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Gantt/multipleDatasetsPanningData';
+import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Gantt/dynamicPanningData';
 
-const graphConfig = utils.deepClone(getGanttPanningConfig('#ganttPanningMultipleDatasets'));
-const dataset = utils.deepClone(graphData);
+/*
+Please refer to the documentation below to see the graphConfig and data objects
+*/
+
+const graphConfig = utils.deepClone(getGanttPanningConfig('#GanttPanningDynamicData'));
+const dataset = [utils.deepClone(exampleData[0])];
 
 const initialState = {
   initial: 0,
   factor: 3,
-  graphConfig,
 };
 
+let graph;
+
 const GanttPanningExample = () => {
+  React.useEffect(() => {
+    graph = Carbon.api.gantt(graphConfig);
+    dataset.forEach((data) => {
+      graph.loadContent(data);
+    });
+  }, []);
+
   const reducer = (panState, action) => {
-    const newGraphConfig = utils.deepClone(panState.graphConfig);
     let hour;
 
     switch (action.type) {
@@ -31,24 +44,29 @@ const GanttPanningExample = () => {
         return panState;
     }
 
-    newGraphConfig.axis.x.lowerLimit = new Date(2016, 0, 1, hour).toISOString();
-    newGraphConfig.axis.x.upperLimit = new Date(2016, 0, 2, hour).toISOString();
+    graph.config.axis.x.lowerLimit = new Date(2016, 0, 1, hour).toISOString();
+    graph.config.axis.x.upperLimit = new Date(2016, 0, 2, hour).toISOString();
+
+    const newDataset = {
+      panData: [utils.deepClone(utils.deepClone(exampleData[1]))],
+    };
+
+    graph.reflowMultipleDatasets(newDataset);
 
     return {
       initial: hour,
       factor: panState.factor,
-      graphConfig: utils.deepClone(newGraphConfig),
     };
   };
 
-  const [panState, dispatch] = React.useReducer(reducer, initialState);
+  const [, dispatch] = React.useReducer(reducer, initialState);
 
   return (
     <React.Fragment>
       <Button className="button-pan-left" text="<" onClick={() => dispatch({ type: 'panLeft' })} />
       <Button className="button-pan-right" text=">" onClick={() => dispatch({ type: 'panRight' })} />
       <div id="tooltip" className="initial-tooltip" />
-      <GanttChart graphID="ganttPanningMultipleDatasets" graphConfig={panState.graphConfig} dataset={dataset} />
+      <div id="GanttPanningDynamicData" />
     </React.Fragment>
   );
 };
