+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/Gantt/Reflow/DynamicEventline.jsx
@@ -1,12 +1,16 @@
 import React from 'react';
-import GanttChart from '@cerner/terra-graphs/lib/components/Gantt/GanttChart';
 import Button from 'terra-button/lib/Button';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/Gantt/GanttChart.module.scss';
 import getGanttPanningConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Gantt/ganttPanning';
 import graphData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Gantt/simplePanningData';
 import Carbon from '@cerner/carbon-graphs/src/js/carbon';
 
+/*
+Please refer to the documentation below to see the graphConfig and data objects
+*/
 const graphConfig = utils.deepClone(getGanttPanningConfig('#GanttPanningWithDynamicEventline'));
 graphConfig.eventline = [
   {
@@ -19,19 +23,23 @@ graphConfig.eventline = [
 ];
 
 const dataset = utils.deepClone(graphData);
+let graph;
 
 const initialState = {
   initial: 0,
   factor: 3,
-  dataset,
-  graphConfig,
 };
 
 const GanttPanningExample = () => {
+  React.useEffect(() => {
+    graph = Carbon.api.gantt(graphConfig);
+    dataset.forEach((data) => {
+      graph.loadContent(data);
+    });
+  }, []);
+
   const reducer = (panState, action) => {
-    const newGraphConfig = utils.deepClone(panState.graphConfig);
     let hour;
-    const newDataset = [utils.deepClone(graphData[0])];
 
     switch (action.type) {
       case 'panLeft':
@@ -44,8 +52,8 @@ const GanttPanningExample = () => {
         return panState;
     }
 
-    newGraphConfig.axis.x.lowerLimit = new Date(2016, 0, 1, hour).toISOString();
-    newGraphConfig.axis.x.upperLimit = new Date(2016, 0, 2, hour).toISOString();
+    graph.config.axis.x.lowerLimit = new Date(2016, 0, 1, hour).toISOString();
+    graph.config.axis.x.upperLimit = new Date(2016, 0, 2, hour).toISOString();
 
     const newEventline = [
       {
@@ -57,25 +65,26 @@ const GanttPanningExample = () => {
       },
     ];
 
+    const newDataset = {
+      eventline: newEventline,
+    };
+
+    graph.reflowMultipleDatasets(newDataset);
+
     return {
       initial: hour,
       factor: panState.factor,
-      dataset: {
-        panData: newDataset,
-        eventline: newEventline,
-      },
-      graphConfig: utils.deepClone(newGraphConfig),
     };
   };
 
-  const [panState, dispatch] = React.useReducer(reducer, initialState);
+  const [, dispatch] = React.useReducer(reducer, initialState);
 
   return (
     <React.Fragment>
       <Button className="button-pan-left" text="<" onClick={() => dispatch({ type: 'panLeft' })} />
       <Button className="button-pan-right" text=">" onClick={() => dispatch({ type: 'panRight' })} />
       <div id="tooltip" className="initial-tooltip" />
-      <GanttChart graphID="GanttPanningWithDynamicEventline" graphConfig={panState.graphConfig} dataset={panState.dataset} />
+      <div id="GanttPanningWithDynamicEventline" />
     </React.Fragment>
   );
 };
