+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/Gantt/BasicGanttCharts/GraphAndLegendPaddingReduced.jsx
@@ -1,17 +1,15 @@
 import React from 'react';
+import GanttChart from '@cerner/terra-graphs/lib/components/Gantt/GanttChart';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
-import '@cerner/terra-graphs/lib/components/Graph.module.scss';
-import '@cerner/terra-graphs/lib/components/Gantt/GanttChart.module.scss';
 import getGanttDefaultConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Gantt/ganttDefault';
 import loadTracks from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Gantt/loadTracks';
 import actionsData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Gantt/actionsData';
-import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
-const graphConfig = utils.deepClone(getGanttDefaultConfig('#ganttGraphAndLegendPaddingReduced'));
+const graphConfig = utils.deepClone(getGanttDefaultConfig('#ganttActions'));
 graphConfig.showActionLegend = true;
 graphConfig.removeContainerPadding = true;
 graphConfig.legendPadding = {
@@ -23,20 +21,9 @@ graphConfig.legendPadding = {
 const actions = utils.deepClone(actionsData);
 const dataset = loadTracks([], [], [], actions, 2);
 
-const GanttExample = () => {
-  React.useEffect(() => {
-    const graph = Carbon.api.gantt(graphConfig);
-    dataset.forEach((data) => {
-      graph.loadContent(data);
-    });
-  }, []);
-
-  return (
-    <React.Fragment>
-      <div id="tooltip" className="initial-tooltip" />
-      <div id="ganttGraphAndLegendPaddingReduced" />
-    </React.Fragment>
-  );
-};
-
-export default GanttExample;
+export default () => (
+  <React.Fragment>
+    <div id="tooltip" className="initial-tooltip" />
+    <GanttChart graphID="ganttActions" graphConfig={graphConfig} dataset={dataset} />
+  </React.Fragment>
+);
