+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/Gantt/Reflow/DynamicData.jsx
@@ -1,37 +1,26 @@
 import React from 'react';
-import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import GanttChart from '@cerner/terra-graphs/lib/components/Gantt/GanttChart';
 import Button from 'terra-button/lib/Button';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
-import '@cerner/terra-graphs/lib/components/Graph.module.scss';
-import '@cerner/terra-graphs/lib/components/Gantt/GanttChart.module.scss';
 import getGanttPanningConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Gantt/ganttPanning';
-import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Gantt/dynamicPanningData';
+import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Gantt/dynamicPanningData';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
-
-const graphConfig = utils.deepClone(getGanttPanningConfig('#GanttPanningDynamicData'));
-const dataset = [utils.deepClone(exampleData[0])];
-
 const initialState = {
   initial: 0,
   factor: 3,
+  dataset: [utils.deepClone(data[0])],
+  graphConfig: utils.deepClone(getGanttPanningConfig('#GanttPanningDynamicData')),
 };
 
-let graph;
-
 const GanttPanningExample = () => {
-  React.useEffect(() => {
-    graph = Carbon.api.gantt(graphConfig);
-    dataset.forEach((data) => {
-      graph.loadContent(data);
-    });
-  }, []);
-
   const reducer = (panState, action) => {
+    const newGraphState = utils.deepClone(panState.graphConfig);
     let hour;
+    const newDataset = [utils.deepClone(data[1])];
 
     switch (action.type) {
       case 'panLeft':
@@ -44,29 +33,25 @@ const GanttPanningExample = () => {
         return panState;
     }
 
-    graph.config.axis.x.lowerLimit = new Date(2016, 0, 1, hour).toISOString();
-    graph.config.axis.x.upperLimit = new Date(2016, 0, 2, hour).toISOString();
-
-    const newDataset = {
-      panData: [utils.deepClone(utils.deepClone(exampleData[1]))],
-    };
-
-    graph.reflowMultipleDatasets(newDataset);
+    newGraphState.axis.x.lowerLimit = new Date(2016, 0, 1, hour).toISOString();
+    newGraphState.axis.x.upperLimit = new Date(2016, 0, 2, hour).toISOString();
 
     return {
       initial: hour,
       factor: panState.factor,
+      dataset: { panData: newDataset },
+      graphConfig: utils.deepClone(newGraphState),
     };
   };
 
-  const [, dispatch] = React.useReducer(reducer, initialState);
+  const [panState, dispatch] = React.useReducer(reducer, initialState);
 
   return (
     <React.Fragment>
       <Button className="button-pan-left" text="<" onClick={() => dispatch({ type: 'panLeft' })} />
       <Button className="button-pan-right" text=">" onClick={() => dispatch({ type: 'panRight' })} />
       <div id="tooltip" className="initial-tooltip" />
-      <div id="GanttPanningDynamicData" />
+      <GanttChart graphID="GanttPanningDynamicData" graphConfig={panState.graphConfig} dataset={panState.dataset} />
     </React.Fragment>
   );
 };
