+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/GanttChart/BasicGanttCharts/LabelTruncation.test.jsx
@@ -1,36 +1,26 @@
 import React from 'react';
+import GanttChart from '@cerner/terra-graphs/lib/components/Gantt/GanttChart';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
-import '@cerner/terra-graphs/lib/components/Graph.module.scss';
-import '@cerner/terra-graphs/lib/components/Gantt/GanttChart.module.scss';
 import getGanttDefaultConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Gantt/ganttDefault';
-import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 
-const graphConfig = utils.deepClone(getGanttDefaultConfig('#ganttLabelTruncation'));
+/*
+Please refer to the documentation below to see the graphConfig and data objects
+*/
+const graphConfig = utils.deepClone(getGanttDefaultConfig('#simpleGantt'));
 graphConfig.showActionLegend = false;
 graphConfig.dateline = [];
 const dataset = [{
   key: 'track 1',
   trackLabel: {
     display:
-      'Really really long project name that cannot be shown realistically',
+    'Really really long project name that cannot be shown realistically',
   },
 }];
 
-const GanttExample = () => {
-  React.useEffect(() => {
-    const graph = Carbon.api.gantt(graphConfig);
-    dataset.forEach((data) => {
-      graph.loadContent(data);
-    });
-  }, []);
-
-  return (
-    <React.Fragment>
-      <div id="tooltip" className="initial-tooltip" />
-      <div id="ganttLabelTruncation" />
-    </React.Fragment>
-  );
-};
-
-export default GanttExample;
+export default () => (
+  <React.Fragment>
+    <div id="tooltip" className="initial-tooltip" />
+    <GanttChart graphID="simpleGantt" graphConfig={graphConfig} dataset={dataset} />
+  </React.Fragment>
+);
