+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/GanttChart/Reflow/SimplePanning.test.jsx
@@ -1,33 +1,24 @@
 import React from 'react';
+import GanttChart from '@cerner/terra-graphs/lib/components/Gantt/GanttChart';
 import Button from 'terra-button/lib/Button';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
-import '@cerner/terra-graphs/lib/components/Graph.module.scss';
-import '@cerner/terra-graphs/lib/components/Gantt/GanttChart.module.scss';
 import getGanttPanningConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Gantt/ganttPanning';
 import graphData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Gantt/simplePanningData';
-import Carbon from '@cerner/carbon-graphs/src/js/carbon';
 
 const graphConfig = utils.deepClone(getGanttPanningConfig('#simpleGanttPanning'));
 
 const dataset = utils.deepClone(graphData);
 
-let graph;
-
 const initialState = {
   initial: 0,
   factor: 3,
+  graphConfig,
 };
 
 const GanttPanningExample = () => {
-  React.useEffect(() => {
-    graph = Carbon.api.gantt(graphConfig);
-    dataset.forEach((data) => {
-      graph.loadContent(data);
-    });
-  }, []);
-
   const reducer = (panState, action) => {
+    const newGraphConfig = utils.deepClone(panState.graphConfig);
     let hour;
 
     switch (action.type) {
@@ -41,25 +32,24 @@ const GanttPanningExample = () => {
         return panState;
     }
 
-    graph.config.axis.x.lowerLimit = new Date(2016, 0, 1, hour).toISOString();
-    graph.config.axis.x.upperLimit = new Date(2016, 0, 2, hour).toISOString();
-
-    graph.reflowMultipleDatasets();
+    newGraphConfig.axis.x.lowerLimit = new Date(2016, 0, 1, hour).toISOString();
+    newGraphConfig.axis.x.upperLimit = new Date(2016, 0, 2, hour).toISOString();
 
     return {
       initial: hour,
       factor: panState.factor,
+      graphConfig: utils.deepClone(newGraphConfig),
     };
   };
 
-  const [, dispatch] = React.useReducer(reducer, initialState);
+  const [panState, dispatch] = React.useReducer(reducer, initialState);
 
   return (
     <React.Fragment>
       <Button className="button-pan-left" text="<" onClick={() => dispatch({ type: 'panLeft' })} />
       <Button className="button-pan-right" text=">" onClick={() => dispatch({ type: 'panRight' })} />
       <div id="tooltip" className="initial-tooltip" />
-      <div id="simpleGanttPanning" />
+      <GanttChart graphID="simpleGanttPanning" graphConfig={panState.graphConfig} dataset={dataset} />
     </React.Fragment>
   );
 };
